% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso.cv.R
\name{glasso_nonconvex_constrained_cv}
\alias{glasso_nonconvex_constrained_cv}
\title{Constrained graphical lasso estimation using the truncated \eqn{\ell_1} penalty.}
\usage{
glasso_nonconvex_constrained_cv(sim, bound = NULL, tau = NULL, num.fold = 5)
}
\arguments{
\item{sim}{A list containing a \eqn{n \times d} data matrix}

\item{bound}{a vector of upper bounds for the constraint}

\item{tau}{tuning parameter for the truncated \eqn{\ell_1} penalty}

\item{num.fold}{fold number for cross validation}
}
\value{
cross validation score.
}
\description{
This function computes the constrained graphical lasso estimator using the truncated \eqn{\ell_1} penalty.
}
