% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso.R
\name{inference_constrained}
\alias{inference_constrained}
\title{Likelihood ratio test for a subset of parameters in Gaussian graphical model}
\usage{
inference_constrained(
  sim,
  para_index,
  bound = NULL,
  tau = NULL,
  inference_type = "LR",
  rho = 1,
  alpha = 1.5,
  eps_abs = 1e-05,
  eps_rel = 1e-06,
  N.iter = 2000,
  dc.iter = 10
)
}
\arguments{
\item{sim}{A list containing a \eqn{n \times d} data matrix and a covariance matrix estimator}

\item{para_index}{indices of the parameters of interest}

\item{bound}{a vector of upper bounds for the constraint}

\item{tau}{tuning parameter for the truncated \eqn{\ell_1} penalty}

\item{inference_type}{either 'LR' or 'LR_gen' with the former based on chi-square test and the later based on normal approximation}
}
\value{
omega.null:

omega.alt:

LR:
}
\description{
This function performs likelihood ratio test for a constrained graphical lasso estimator using the truncated \eqn{\ell_1} penalty.
}
